+++ src/BetterSafe.cpp
@@ -1,7 +1,4 @@
 #include "BetterSafe.hpp"
-#include <Geode/binding/FLAlertLayer.hpp>
-#include <Geode/binding/GJGameLevel.hpp>
-#include <Geode/binding/LoadingCircle.hpp>
 #include <Geode/utils/ranges.hpp>
 #include <Geode/utils/string.hpp>
 
@@ -11,35 +8,12 @@ using namespace geode::prelude;
 #define WEEKLY_SAFE_URL "https://raw.githubusercontent.com/hiimjasmine00/the-safe/master/v2/weekly.json"
 #define EVENT_SAFE_URL "https://raw.githubusercontent.com/hiimjasmine00/the-safe/master/v2/event.json"
 
-SafeDate BetterSafe::parseDate(const std::string& date) {
-    auto parts = string::split(date, "-");
-    return {
-        .year = parts.size() > 0 ? numFromString<int>(parts[0]).unwrapOr(1970) : 1970,
-        .month = parts.size() > 1 ? numFromString<int>(parts[1]).unwrapOr(1) : 1,
-        .day = parts.size() > 2 ? numFromString<int>(parts[2]).unwrapOr(1) : 1
-    };
-}
-
-SafeDate BetterSafe::dateFromTime(time_t time) {
-    #pragma clang diagnostic push
-    #pragma clang diagnostic ignored "-Wdeprecated-declarations"
-    auto timeinfo = std::localtime(&time);
-    #pragma clang diagnostic pop
-    return { timeinfo->tm_year + 1900, timeinfo->tm_mon + 1, timeinfo->tm_mday };
-}
-
-void BetterSafe::loadSafe(
-    GJTimedLevelType type, EventListener<web::WebTask>&& listenerRef, const std::function<void()>& success, const std::function<void(int)>& failure
-) {
+void BetterSafe::loadSafe(GJTimedLevelType type, EventListener<web::WebTask>* listener, std::function<void()> success, std::function<void(int)> failure) {
     if (!getSafeLevels(type).empty()) return success();
 
-    auto&& listener = std::move(listenerRef);
-
-    listener.bind([failure, success, type](web::WebTask::Event* e) {
+    listener->bind([failure = std::move(failure), success = std::move(success), type](web::WebTask::Event* e) {
         if (auto res = e->getValue()) {
-            if (!res->ok()) {
-                return failure(res->code());
-            }
+            if (!res->ok()) return failure(res->code());
 
             auto json = res->json().unwrapOr(std::vector<matjson::Value>());
             if (!json.isArray()) return success();
@@ -49,7 +23,14 @@ void BetterSafe::loadSafe(
                     .id = (int)level["id"].asInt().unwrapOr(0),
                     .timelyID = (int)level["timelyID"].asInt().unwrapOr(0),
                     .dates = level.contains("dates") && level["dates"].isArray() ? ranges::map<std::vector<SafeDate>>(level["dates"].asArray().unwrap(),
-                        [](const matjson::Value& date) { return parseDate(date.asString().unwrapOr("1970-01-01")); }) : std::vector<SafeDate>(),
+                        [](const matjson::Value& date) {
+                            auto parts = string::split(date.asString().unwrapOr("1970-01-01"), "-");
+                            return SafeDate {
+                                .year = parts.size() > 0 ? numFromString<int>(parts[0]).unwrapOr(1970) : 1970,
+                                .month = parts.size() > 1 ? numFromString<int>(parts[1]).unwrapOr(1) : 1,
+                                .day = parts.size() > 2 ? numFromString<int>(parts[2]).unwrapOr(1) : 1
+                            };
+                        }) : std::vector<SafeDate>(),
                     .type = type,
                     .tier = (int)level["tier"].asInt().unwrapOr(0)
                 };
@@ -57,32 +38,46 @@ void BetterSafe::loadSafe(
 
             if (type != GJTimedLevelType::Event || EVENT_SAFE.size() < 2) return success();
 
-            auto lastEventDate = EVENT_SAFE[1].dates.back();
-            tm timeinfo = { 0, 0, 0, lastEventDate.day, lastEventDate.month - 1, lastEventDate.year - 1900 };
-            auto currentDate = dateFromTime(time(0));
-            tm currentTimeinfo = { 0, 0, 0, currentDate.day, currentDate.month - 1, currentDate.year - 1900 };
-            auto truncatedTime = mktime(&currentTimeinfo);
-            for (auto lastEventTime = mktime(&timeinfo) + 86400; lastEventTime <= truncatedTime; lastEventTime += 86400) {
-                EVENT_SAFE[0].dates.push_back(dateFromTime(lastEventTime));
+            auto currentTime = time(0);
+            #pragma clang diagnostic push
+            #pragma clang diagnostic ignored "-Wdeprecated-declarations"
+            auto currentTimeinfo = std::localtime(&currentTime);
+            #pragma clang diagnostic pop
+            currentTimeinfo->tm_hour = 0;
+            currentTimeinfo->tm_min = 0;
+            currentTimeinfo->tm_sec = 0;
+            auto& [lastYear, lastMonth, lastDay] = EVENT_SAFE[1].dates.back();
+            tm timeinfo = { 0, 0, 0, lastDay, lastMonth - 1, lastYear - 1900 };
+            auto truncatedTime = mktime(currentTimeinfo);
+            for (auto lastTime = mktime(&timeinfo) + 86400; lastTime <= truncatedTime; lastTime += 86400) {
+                #pragma clang diagnostic push
+                #pragma clang diagnostic ignored "-Wdeprecated-declarations"
+                auto timeinfo = std::localtime(&lastTime);
+                #pragma clang diagnostic pop
+                EVENT_SAFE[0].dates.push_back({ timeinfo->tm_year + 1900, timeinfo->tm_mon + 1, timeinfo->tm_mday });
             }
 
             success();
         }
     });
 
+    auto url = "";
     switch (type) {
-        case GJTimedLevelType::Daily: listener.setFilter(web::WebRequest().get(DAILY_SAFE_URL)); break;
-        case GJTimedLevelType::Weekly: listener.setFilter(web::WebRequest().get(WEEKLY_SAFE_URL)); break;
-        case GJTimedLevelType::Event: listener.setFilter(web::WebRequest().get(EVENT_SAFE_URL)); break;
-        default: break;
+        case GJTimedLevelType::Daily: url = DAILY_SAFE_URL; break;
+        case GJTimedLevelType::Weekly: url = WEEKLY_SAFE_URL; break;
+        case GJTimedLevelType::Event: url = EVENT_SAFE_URL; break;
+        default: return;
     }
+    listener->setFilter(web::WebRequest().get(url));
 }
 
 std::vector<SafeLevel> BetterSafe::getMonth(int year, int month, GJTimedLevelType type) {
-    return ranges::reduce<std::vector<SafeLevel>>(getSafeLevels(type), [year, month](std::vector<SafeLevel>& levels, const SafeLevel& level) {
-        ranges::push(levels, ranges::map<std::vector<SafeLevel>>(ranges::filter<std::vector<SafeDate>>(level.dates, [year, month](const SafeDate& date) {
-            return date.year == year && date.month == month;
-        }), [&level](const auto&) { return level; }));
+    using SafeLevels = std::vector<SafeLevel>;
+
+    return ranges::reduce<SafeLevels>(getSafeLevels(type), [year, month](SafeLevels& levels, const SafeLevel& level) {
+        ranges::push(levels, ranges::reduce<SafeLevels>(level.dates, [&level, year, month](SafeLevels& dates, const SafeDate& date) {
+            if (date.year == year && date.month == month) dates.push_back(level);
+        }));
     });
 }
 
@@ -96,11 +91,4 @@ std::vector<SafeLevel>& BetterSafe::getSafeLevels(GJTimedLevelType type) {
             return empty;
         }
     }
-}
-
-int BetterSafe::getDifficultyFromLevel(GJGameLevel* level) {
-    if (level->m_demon > 0) return level->m_demonDifficulty > 0 ? level->m_demonDifficulty + 4 : 6;
-    else if (level->m_autoLevel) return -1;
-    else if (level->m_ratings < 5) return 0;
-    else return level->m_ratingsSum / level->m_ratings;
 }