+++ src/BetterSafe.hpp
@@ -22,10 +22,7 @@ public:
     inline static std::vector<SafeLevel> WEEKLY_SAFE;
     inline static std::vector<SafeLevel> EVENT_SAFE;
 
-    static SafeDate parseDate(const std::string&);
-    static SafeDate dateFromTime(time_t);
-    static void loadSafe(GJTimedLevelType, geode::EventListener<geode::utils::web::WebTask>&&, const std::function<void()>&, const std::function<void(int)>&);
+    static void loadSafe(GJTimedLevelType, geode::EventListener<geode::utils::web::WebTask>*, std::function<void()>, std::function<void(int)>);
     static std::vector<SafeLevel> getMonth(int year, int month, GJTimedLevelType type);
     static std::vector<SafeLevel>& getSafeLevels(GJTimedLevelType type);
-    static int getDifficultyFromLevel(GJGameLevel* level);
 };