+++ src/classes/BSCalendarPopup.cpp
@@ -1,6 +1,7 @@
 #include "BSCalendarPopup.hpp"
 #include "BSSelectPopup.hpp"
 #include <Geode/binding/GameLevelManager.hpp>
+#include <Geode/binding/GameToolbox.hpp>
 #include <Geode/binding/GJGameLevel.hpp>
 #include <Geode/binding/GJSearchObject.hpp>
 #include <Geode/binding/LoadingCircle.hpp>
@@ -40,17 +41,9 @@ bool BSCalendarPopup::setup(CCObject* obj, SEL_MenuHandler onSafe, GJTimedLevelT
     createWeekdayLabel("Sat", (int)sundayFirst + 5);
     createWeekdayLabel("Sun", sundayFirst ? 0 : 6);
 
-    GameLevelManager::get()->m_levelManagerDelegate = this;
-
     m_prevButton = CCMenuItemExt::createSpriteExtraWithFrameName("GJ_arrow_01_001.png", 1.0f, [this](auto) {
-        if (m_year == m_firstYear && m_month == m_firstMonth) return;
-
-        if (m_month == 1) {
-            m_year--;
-            m_month = 12;
-        }
-        else m_month--;
-        loadMonth();
+        if (m_year > m_minYear || (m_year == m_minYear && m_month > m_minMonth))
+            loadMonth(m_month == 1 ? m_year - 1 : m_year, m_month == 1 ? 12 : m_month - 1);
     });
     m_prevButton->setPosition({ -34.5f, 140.0f });
     m_prevButton->setVisible(false);
@@ -60,13 +53,8 @@ bool BSCalendarPopup::setup(CCObject* obj, SEL_MenuHandler onSafe, GJTimedLevelT
     auto nextButtonSprite = CCSprite::createWithSpriteFrameName("GJ_arrow_01_001.png");
     nextButtonSprite->setFlipX(true);
     m_nextButton = CCMenuItemExt::createSpriteExtra(nextButtonSprite, [this](auto) {
-        if (m_year == m_currentYear && m_month == m_currentMonth) return;
-        if (m_month == 12) {
-            m_year++;
-            m_month = 1;
-        }
-        else m_month++;
-        loadMonth();
+        if (m_year < m_maxYear || (m_year == m_maxYear && m_month < m_maxMonth))
+            loadMonth(m_month == 12 ? m_year + 1 : m_year, m_month == 12 ? 1 : m_month + 1);
     });
     m_nextButton->setPosition({ 334.5f, 140.0f });
     m_nextButton->setVisible(false);
@@ -75,13 +63,11 @@ bool BSCalendarPopup::setup(CCObject* obj, SEL_MenuHandler onSafe, GJTimedLevelT
 
     auto firstArrow = CCSprite::createWithSpriteFrameName("GJ_arrow_02_001.png");
     auto otherFirstArrow = CCSprite::createWithSpriteFrameName("GJ_arrow_02_001.png");
-    otherFirstArrow->setPosition(firstArrow->getContentSize() / 2 - CCPoint { 20.0f, 0.0f });
+    otherFirstArrow->setPosition(firstArrow->getContentSize() / 2.0f - CCPoint { 20.0f, 0.0f });
     firstArrow->addChild(otherFirstArrow);
     firstArrow->setScale(0.4f);
     m_firstButton = CCMenuItemExt::createSpriteExtra(firstArrow, [this](auto) {
-        m_year = m_firstYear;
-        m_month = m_firstMonth;
-        loadMonth();
+        if (m_year > m_minYear || (m_year == m_minYear && m_month > m_minMonth)) loadMonth(m_minYear, m_minMonth);
     });
     m_firstButton->setPosition({ -27.5f, 175.0f });
     m_firstButton->setVisible(false);
@@ -91,14 +77,12 @@ bool BSCalendarPopup::setup(CCObject* obj, SEL_MenuHandler onSafe, GJTimedLevelT
     auto lastArrow = CCSprite::createWithSpriteFrameName("GJ_arrow_02_001.png");
     lastArrow->setFlipX(true);
     auto otherLastArrow = CCSprite::createWithSpriteFrameName("GJ_arrow_02_001.png");
-    otherLastArrow->setPosition(lastArrow->getContentSize() / 2 + CCPoint { 20.0f, 0.0f });
+    otherLastArrow->setPosition(lastArrow->getContentSize() / 2.0f + CCPoint { 20.0f, 0.0f });
     otherLastArrow->setFlipX(true);
     lastArrow->addChild(otherLastArrow);
     lastArrow->setScale(0.4f);
     m_lastButton = CCMenuItemExt::createSpriteExtra(lastArrow, [this](auto) {
-        m_year = m_currentYear;
-        m_month = m_currentMonth;
-        loadMonth();
+        if (m_year < m_maxYear || (m_year == m_maxYear && m_month < m_maxMonth)) loadMonth(m_maxYear, m_maxMonth);
     });
     m_lastButton->setPosition({ 327.5f, 175.0f });
     m_lastButton->setVisible(false);
@@ -115,11 +99,7 @@ bool BSCalendarPopup::setup(CCObject* obj, SEL_MenuHandler onSafe, GJTimedLevelT
     m_monthLabel = CCLabelBMFont::create("", "goldFont.fnt");
     m_monthLabel->setScale(0.6f);
     m_monthButton = CCMenuItemExt::createSpriteExtra(m_monthLabel, [this](auto) {
-        BSSelectPopup::create(m_year, m_month, m_firstYear, m_firstMonth, m_currentYear, m_currentMonth, [this](int year, int month) {
-            m_year = year;
-            m_month = month;
-            loadMonth();
-        })->show();
+        BSSelectPopup::create(this, [this](int year, int month) { loadMonth(year, month); })->show();
     });
     m_monthButton->setPosition({ 150.0f, 265.0f });
     m_monthButton->setEnabled(false);
@@ -127,12 +107,11 @@ bool BSCalendarPopup::setup(CCObject* obj, SEL_MenuHandler onSafe, GJTimedLevelT
     m_buttonMenu->addChild(m_monthButton);
 
     m_loadingCircle = LoadingCircle::create();
-    m_loadingCircle->setParentLayer(m_mainLayer);
     m_loadingCircle->setContentSize({ 300.0f, 280.0f });
     m_loadingCircle->m_sprite->setPosition({ 150.0f, 140.0f });
-    m_loadingCircle->retain();
-    m_loadingCircle->show();
+    m_loadingCircle->setParentLayer(m_mainLayer);
     m_loadingCircle->setID("loading-circle");
+    m_loadingCircle->show();
 
     auto safeButton = CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("GJ_safeBtn_001.png"), obj, onSafe);
     safeButton->setPosition({ 340.0f, 25.0f });
@@ -140,7 +119,7 @@ bool BSCalendarPopup::setup(CCObject* obj, SEL_MenuHandler onSafe, GJTimedLevelT
     safeButton->setID("safe-button");
     m_buttonMenu->addChild(safeButton);
 
-    auto refreshButton = CCMenuItemExt::createSpriteExtraWithFrameName("GJ_updateBtn_001.png", 1.0f, [this, type](auto) {
+    auto refreshButton = CCMenuItemExt::createSpriteExtraWithFrameName("GJ_updateBtn_001.png", 1.0f, [this](auto) {
         m_loadingCircle->setVisible(true);
         if (m_hoverNode) m_hoverNode->close();
         m_calendarMenu->removeAllChildren();
@@ -149,46 +128,15 @@ bool BSCalendarPopup::setup(CCObject* obj, SEL_MenuHandler onSafe, GJTimedLevelT
         m_firstButton->setVisible(false);
         m_lastButton->setVisible(false);
         m_monthButton->setEnabled(false);
-        BetterSafe::getSafeLevels(type).clear();
-        BetterSafe::loadSafe(type, std::move(m_listener), [this, type] {
-            if (!BetterSafe::getSafeLevels(type).empty()) loadMonth();
-        }, [this](int code) {
-            FLAlertLayer::create(fmt::format("Load Failed ({})", code).c_str(), "Failed to load safe levels. Please try again later.", "OK")->show();
-            m_loadingCircle->setVisible(false);
-        });
+        BetterSafe::getSafeLevels(m_type).clear();
+        loadSafe(true);
     });
     refreshButton->setPosition({ 340.0f, 80.0f });
     refreshButton->setID("refresh-button");
     m_buttonMenu->addChild(refreshButton);
 
-    auto dateNow = BetterSafe::dateFromTime(time(0));
-    m_year = dateNow.year;
-    m_month = dateNow.month;
-    BetterSafe::loadSafe(type, std::move(m_listener), [this, type] {
-        auto& safe = BetterSafe::getSafeLevels(type);
-        if (safe.empty()) return;
-
-        auto firstLevel = safe[0];
-        auto lastLevel = safe.back();
-        auto firstDate = ranges::find(firstLevel.dates, [this](const SafeDate& date) { return date.year == m_year && date.month == m_month; });
-        if (!firstDate.has_value()) {
-            auto lastDate = !firstLevel.dates.empty() ? firstLevel.dates.back() : SafeDate { m_year, m_month, 1 };
-            m_currentYear = lastDate.year;
-            m_currentMonth = lastDate.month;
-        } else {
-            m_currentYear = firstDate->year;
-            m_currentMonth = firstDate->month;
-        }
-        auto startDate = !lastLevel.dates.empty() ? lastLevel.dates[0] : SafeDate { m_year, m_month, 1 };
-        m_firstYear = startDate.year;
-        m_firstMonth = startDate.month;
-        loadMonth();
-    }, [this](int code) {
-        FLAlertLayer::create(fmt::format("Load Failed ({})", code).c_str(), "Failed to load safe levels. Please try again later.", "OK")->show();
-        m_loadingCircle->setVisible(false);
-    });
-
     handleTouchPriority(this);
+    loadSafe();
 
     return true;
 }
@@ -201,56 +149,87 @@ void BSCalendarPopup::createWeekdayLabel(const char* text, int idx) {
     m_mainLayer->addChild(label);
 }
 
-void BSCalendarPopup::loadMonth() {
-    m_loadingCircle->setVisible(true);
+void BSCalendarPopup::loadSafe(bool refresh) {
+    BetterSafe::loadSafe(m_type, &m_listener, [this, refresh] {
+        auto& safe = BetterSafe::getSafeLevels(m_type);
+        if (safe.empty()) return;
+
+        #pragma clang diagnostic push
+        #pragma clang diagnostic ignored "-Wdeprecated-declarations"
+        auto currentTime = time(0);
+        auto timeinfo = std::localtime(&currentTime);
+        #pragma clang diagnostic pop
+        auto currentYear = timeinfo->tm_year + 1900;
+        auto currentMonth = timeinfo->tm_mon + 1;
+
+        auto& frontDates = safe.front().dates;
+        auto maxDate = frontDates.end() - (int)frontDates.empty();
+        m_maxYear = maxDate != frontDates.end() ? maxDate->year : currentYear;
+        m_maxMonth = maxDate != frontDates.end() ? maxDate->month : currentMonth;
+
+        auto& backDates = safe.back().dates;
+        auto minDate = backDates.begin();
+        m_minYear = minDate != backDates.end() ? minDate->year : currentYear;
+        m_minMonth = minDate != backDates.end() ? minDate->month : currentMonth;
+
+        loadMonth(m_year == 0 ? currentYear : m_year, m_month == 0 ? currentMonth : m_month, refresh);
+    }, [this](int code) {
+        FLAlertLayer::create(fmt::format("Load Failed ({})", code).c_str(), "Failed to load safe levels. Please try again later.", "OK")->show();
+        m_loadingCircle->setVisible(false);
+    });
+}
+
+void BSCalendarPopup::loadMonth(int year, int month, bool refresh) {
+    m_year = year;
+    m_month = month;
 
     if (m_hoverNode) m_hoverNode->close();
-
     m_calendarMenu->removeAllChildren();
-    m_monthLabel->setString(fmt::format("{} {}", MONTHS[m_month - 1], m_year).c_str());
+    m_monthLabel->setString(fmt::format("{} {}", BSSelectPopup::MONTHS[m_month - 1], m_year).c_str());
     m_monthButton->updateSprite();
     m_monthButton->setEnabled(false);
-
     m_prevButton->setVisible(false);
     m_nextButton->setVisible(false);
     m_firstButton->setVisible(false);
     m_lastButton->setVisible(false);
 
     auto levelSafe = BetterSafe::getMonth(m_year, m_month, m_type);
-    if (levelSafe.empty()) return loadLevelsFinished(CCArray::create(), "");
+    if (levelSafe.empty()) return setupMonth(nullptr);
 
-    auto searchObject = GJSearchObject::create(SearchType::MapPackOnClick, string::join(ranges::map<std::vector<std::string>>(levelSafe, [](const SafeLevel& level) {
-        return std::to_string(level.id);
-    }), ","));
+    auto searchObject = GJSearchObject::create(SearchType::MapPackOnClick, ranges::reduce<std::string>(levelSafe,
+        [](std::string& str, const SafeLevel& level) { str += (str.empty() ? "" : ",") + std::to_string(level.id); }));
+
     auto glm = GameLevelManager::get();
-    std::string key = searchObject->getKey();
-    if (auto storedLevels = glm->getStoredOnlineLevels(key.substr(std::max(0, (int)key.size() - 256)).c_str()))
-        loadLevelsFinished(storedLevels, searchObject->getKey());
-    else glm->getOnlineLevels(searchObject);
+    if (!refresh) {
+        std::string key = searchObject->getKey();
+        if (auto storedLevels = glm->getStoredOnlineLevels(key.substr(std::max(0, (int)key.size() - 256)).c_str())) return setupMonth(storedLevels);
+    }
+
+    m_loadingCircle->setVisible(true);
+    glm->m_levelManagerDelegate = this;
+    glm->getOnlineLevels(searchObject);
 }
 
-void BSCalendarPopup::loadLevelsFinished(CCArray* levels, const char*) {
-    CC_SAFE_RELEASE(m_levels);
-    CC_SAFE_RETAIN(levels);
-    m_levels = levels;
-    setupMonth();
+void BSCalendarPopup::loadLevelsFinished(CCArray* levels, const char*, int) {
+    setupMonth(levels);
 }
 
-void BSCalendarPopup::loadLevelsFailed(const char*) {
+void BSCalendarPopup::loadLevelsFailed(const char*, int) {
     m_loadingCircle->setVisible(false);
     FLAlertLayer::create("Load Failed", "Failed to load safe levels. Please try again later.", "OK")->show();
 }
 
-void BSCalendarPopup::setupMonth() {
+void BSCalendarPopup::setupMonth(CCArray* levels) {
     m_monthButton->setEnabled(true);
     m_loadingCircle->setVisible(false);
 
-    m_prevButton->setVisible(m_year != m_firstYear || m_month != m_firstMonth);
-    m_firstButton->setVisible(m_year != m_firstYear || m_month != m_firstMonth);
-    m_nextButton->setVisible(m_year != m_currentYear || m_month != m_currentMonth);
-    m_lastButton->setVisible(m_year != m_currentYear || m_month != m_currentMonth);
+    m_prevButton->setVisible(m_year > m_minYear || (m_year == m_minYear && m_month > m_minMonth));
+    m_firstButton->setVisible(m_year > m_minYear || (m_year == m_minYear && m_month > m_minMonth));
+    m_nextButton->setVisible(m_year < m_maxYear || (m_year == m_maxYear && m_month < m_maxMonth));
+    m_lastButton->setVisible(m_year < m_maxYear || (m_year == m_maxYear && m_month < m_maxMonth));
 
-    auto daysInMonth = m_month == 2 && m_year % 4 == 0 && (m_year % 100 != 0 || m_year % 400 == 0) ? 29 : DAYS_IN_MONTH[m_month - 1];
+    if (!levels) return;
+
     tm timeinfo = { 0, 0, 0, 1, m_month - 1, m_year - 1900 };
     auto time = mktime(&timeinfo);
     #pragma clang diagnostic push
@@ -260,70 +239,126 @@ void BSCalendarPopup::setupMonth() {
     if (!Mod::get()->getSettingValue<bool>("sunday-first")) firstWeekday = (firstWeekday + 6) % 7;
 
     auto levelSafe = BetterSafe::getMonth(m_year, m_month, m_type);
+    auto daysInMonth = 0;
+    switch (m_month) {
+        case 1: case 3: case 5: case 7: case 8: case 10: case 12: daysInMonth = 31; break;
+        case 2: daysInMonth = 28 + (m_year % 4 == 0 && (m_year % 100 != 0 || m_year % 400 == 0)); break;
+        case 4: case 6: case 9: case 11: daysInMonth = 30; break;
+    }
     for (int i = 0; i < daysInMonth; i++) {
-        auto safeLevelOpt = ranges::find(levelSafe, [this, i](const SafeLevel& level) {
-            return ranges::contains(level.dates, [this, i](const SafeDate& date) {
+        auto safeLevelIt = std::ranges::find_if(levelSafe, [this, i](const SafeLevel& level) {
+            return std::ranges::any_of(level.dates, [this, i](const SafeDate& date) {
                 return date.year == m_year && date.month == m_month && date.day == i + 1;
             });
         });
-        if (!safeLevelOpt.has_value()) continue;
-        auto safeLevel = safeLevelOpt.value();
-        auto gameLevelOpt =  ranges::find(CCArrayExt<GJGameLevel*>(m_levels), [safeLevel](GJGameLevel* level) {
-            return level->m_levelID == safeLevel.id;
-        });
-        if (!gameLevelOpt.has_value()) continue;
-        auto gameLevel = gameLevelOpt.value();
+        if (safeLevelIt == levelSafe.end()) continue;
 
-        auto levelDifficulty = BetterSafe::getDifficultyFromLevel(gameLevel);
+        auto& safeLevel = *safeLevelIt;
+        auto gameLevels = reinterpret_cast<GJGameLevel**>(levels->data->arr);
+        auto gameLevelsEnd = gameLevels + levels->count();
+        auto gameLevelIt = std::find_if(gameLevels, gameLevelsEnd, [&safeLevel](GJGameLevel* level) {
+            return level->m_levelID.value() == safeLevel.id;
+        });
+        if (gameLevelIt == gameLevelsEnd) continue;
+
+        auto gameLevel = *gameLevelIt;
+
+        auto levelDifficulty = gameLevel->m_demon > 0 ? gameLevel->m_demonDifficulty > 0 ? gameLevel->m_demonDifficulty + 4 : 6 :
+            gameLevel->m_autoLevel ? -1 : gameLevel->m_ratings < 5 ? 0 : gameLevel->m_ratingsSum / gameLevel->m_ratings;
         auto diffFrame = levelDifficulty == -1 ? "diffIcon_auto_btn_001.png" : fmt::format("diffIcon_{:02d}_btn_001.png", levelDifficulty);
         if (auto moreDifficulties = Loader::get()->getLoadedMod("uproxide.more_difficulties")) {
             auto legacy = moreDifficulties->getSettingValue<bool>("legacy-difficulties");
             auto levelStars = gameLevel->m_stars.value();
             if (moreDifficulties->getSavedValue<bool>("casual", true) && levelDifficulty == 3 && levelStars == 4)
-                diffFrame = legacy ? "uproxide.more_difficulties/MD_Difficulty04Small_Legacy.png" : "uproxide.more_difficulties/MD_Difficulty04Small.png";
+                diffFrame = fmt::format("uproxide.more_difficulties/MD_Difficulty04Small{}.png", legacy ? "_Legacy" : "");
             else if (moreDifficulties->getSavedValue<bool>("tough", true) && levelDifficulty == 4 && levelStars == 7)
-                diffFrame = legacy ? "uproxide.more_difficulties/MD_Difficulty07Small_Legacy.png" : "uproxide.more_difficulties/MD_Difficulty07Small.png";
+                diffFrame = fmt::format("uproxide.more_difficulties/MD_Difficulty07Small{}.png", legacy ? "_Legacy" : "");
             else if (moreDifficulties->getSavedValue<bool>("cruel", true) && levelDifficulty == 5 && levelStars == 9)
-                diffFrame = legacy ? "uproxide.more_difficulties/MD_Difficulty09Small_Legacy.png" : "uproxide.more_difficulties/MD_Difficulty09Small.png";
+                diffFrame = fmt::format("uproxide.more_difficulties/MD_Difficulty09Small{}.png", legacy ? "_Legacy" : "");
         }
+
         auto hasBetweenDemon = false;
         if (auto demonsInBetween = Loader::get()->getLoadedMod("hiimjustin000.demons_in_between")) {
             if (demonsInBetween->getSettingValue<bool>("enable-difficulties") && safeLevel.tier > 0) {
-                diffFrame = fmt::format("hiimjustin000.demons_in_between/DIB_{:02d}_001.png", DIFFICULTY_INDICES[safeLevel.tier]);
-                hasBetweenDemon = true;
+                auto betweenDifficulty = 0;
+                switch (safeLevel.tier) {
+                    case 1: betweenDifficulty = 1; break;
+                    case 2: betweenDifficulty = 2; break;
+                    case 3: betweenDifficulty = 3; break;
+                    case 4: betweenDifficulty = 4; break;
+                    case 5: betweenDifficulty = 5; break;
+                    case 6: betweenDifficulty = 6; break;
+                    case 7: betweenDifficulty = 7; break;
+                    case 8: betweenDifficulty = 8; break;
+                    case 9: betweenDifficulty = 9; break;
+                    case 10: betweenDifficulty = 10; break;
+                    case 11: case 12: betweenDifficulty = 11; break;
+                    case 13: betweenDifficulty = 12; break;
+                    case 14: betweenDifficulty = 13; break;
+                    case 15: case 16: betweenDifficulty = 14; break;
+                    case 17: case 18: betweenDifficulty = 15; break;
+                    case 19: betweenDifficulty = 16; break;
+                    case 20: betweenDifficulty = 17; break;
+                    case 21: betweenDifficulty = 18; break;
+                    case 22: betweenDifficulty = 19; break;
+                    case 23: case 24: case 25: case 26: case 27: case 28: case 29:
+                    case 30: case 31: case 32: case 33: case 34: case 35: betweenDifficulty = 20; break;
+                }
+                if (betweenDifficulty > 0) {
+                    diffFrame = fmt::format("hiimjustin000.demons_in_between/DIB_{:02d}_001.png", betweenDifficulty);
+                    hasBetweenDemon = true;
+                }
             }
         }
+
         auto diffIcon = CCSprite::createWithSpriteFrameName(diffFrame.c_str());
         diffIcon->setScale(0.75f);
-        auto featureFrame = "";
-        auto levelFeature = gameLevel->m_featured > 0 ? gameLevel->m_isEpic + 1 : 0;
-        switch (levelFeature) {
-            case 1: featureFrame = "GJ_featuredCoin_001.png"; break;
-            case 2: featureFrame = "GJ_epicCoin_001.png"; break;
-            case 3: featureFrame = "GJ_epicCoin2_001.png"; break;
-            case 4: featureFrame = "GJ_epicCoin3_001.png"; break;
+
+        if (gameLevel->m_featured > 0) {
+            auto featureFrame = "";
+            switch (gameLevel->m_isEpic) {
+                case 0: featureFrame = "GJ_featuredCoin_001.png"; break;
+                case 1: featureFrame = "GJ_epicCoin_001.png"; break;
+                case 2: featureFrame = "GJ_epicCoin2_001.png"; break;
+                case 3: featureFrame = "GJ_epicCoin3_001.png"; break;
+            }
+            if (auto featureIcon = CCSprite::createWithSpriteFrameName(featureFrame)) {
+                featureIcon->setPosition(diffIcon->getContentSize() / 2.0f + CCPoint { 0.0f, -5.5f });
+                if (hasBetweenDemon) featureIcon->setPositionY(9.5f);
+                diffIcon->addChild(featureIcon, -2);
+                if (gameLevel->m_isEpic == 3) {
+                    auto mythicParticles = GameToolbox::particleFromString(
+                        "30a-1a1.3a0.2a20a90a0a10a5a20a20a0a0a8a0a0a0a4a1a0a0a0a0a1a0a1a0a1a0a1a1a0a0a0a0a0.784314"
+                        "a0a1a0a1a0a0.27a0a0.27a0a0a0a0a0a0a0a0a2a1a0a0a0a0a0a0a0.25a0a0a0a0a0a0a0a0a0a0a0",
+                        nullptr, false
+                    );
+                    mythicParticles->setPosition(featureIcon->getPosition());
+                    mythicParticles->setScale(0.9f);
+                    mythicParticles->resetSystem();
+                    diffIcon->addChild(mythicParticles, -1);
+                    for (int i = 0; i < 5; i++) {
+                        mythicParticles->update(0.15f);
+                    }
+                }
+            }
         }
-        if (levelFeature > 0) {
-            auto featureIcon = CCSprite::createWithSpriteFrameName(featureFrame);
-            featureIcon->setPosition(diffIcon->getContentSize() / 2 + CCPoint { 0.0f, -5.5f });
-            if (hasBetweenDemon) featureIcon->setPositionY(9.5f);
-            diffIcon->addChild(featureIcon, -2);
-        }
-        auto hoverButton = CCMenuItemExt::createSpriteExtra(diffIcon, [this, gameLevel, safeLevel](auto sender) {
+
+        auto hoverButton = CCMenuItemExt::createSpriteExtra(diffIcon, [this, gameLevel, safeLevel](CCMenuItemSpriteExtra* sender) {
             if (m_hoverNode) m_hoverNode->close();
             m_hoverNode = BSHoverNode::create(safeLevel, gameLevel, [this] { m_hoverNode = nullptr; });
-            m_hoverNode->setPosition(sender->getPosition() + CCPoint { 0.0f, sender->getContentHeight() / 2 + m_hoverNode->getContentHeight() / 2 + 5.0f });
+            m_hoverNode->setPosition(sender->getPosition() + CCPoint {
+                0.0f,
+                sender->getContentHeight() / 2.0f + m_hoverNode->getContentHeight() / 2.0f + 5.0f
+            });
             m_mainLayer->addChild(m_hoverNode, 200);
         });
-        hoverButton->setPosition({ (i + firstWeekday) % 7 * 38.0f + 36.0f, 219.0f - floorf(((float)i + firstWeekday) / 7) * 38.0f });
+        hoverButton->setPosition({ (i + firstWeekday) % 7 * 38.0f + 36.0f, 219.0f - (int)((i + firstWeekday) / 7) * 38.0f });
         hoverButton->setID(fmt::format("level-button-{}", i + 1));
         m_calendarMenu->addChild(hoverButton);
     }
 }
 
 BSCalendarPopup::~BSCalendarPopup() {
-    CC_SAFE_RELEASE(m_levels);
-    CC_SAFE_RELEASE(m_loadingCircle);
     auto glm = GameLevelManager::get();
     if (glm->m_levelManagerDelegate == this) glm->m_levelManagerDelegate = nullptr;
 }