+++ src/classes/BSCalendarPopup.hpp
@@ -3,25 +3,15 @@
 #include <Geode/ui/Popup.hpp>
 
 class BSCalendarPopup : public geode::Popup<cocos2d::CCObject*, cocos2d::SEL_MenuHandler, GJTimedLevelType>, public LevelManagerDelegate {
-public:
-    inline static std::vector<int> DAYS_IN_MONTH = { 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31 };
-    inline static std::vector<std::string> MONTHS = {
-        "January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"
-    };
-    inline static std::vector<int> DIFFICULTY_INDICES = {
-        0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 11, 12, 13, 14, 14, 15, 15,
-        16, 17, 18, 19, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20, 20
-    };
 protected:
     geode::EventListener<geode::utils::web::WebTask> m_listener;
     GJTimedLevelType m_type;
     int m_year;
-    int m_firstYear;
-    int m_currentYear;
+    int m_minYear;
+    int m_maxYear;
     int m_month;
-    int m_firstMonth;
-    int m_currentMonth;
-    cocos2d::CCArray* m_levels;
+    int m_minMonth;
+    int m_maxMonth;
     cocos2d::CCMenu* m_calendarMenu;
     cocos2d::CCLabelBMFont* m_monthLabel;
     CCMenuItemSpriteExtra* m_monthButton;
@@ -34,21 +24,23 @@ protected:
 
     bool setup(CCObject*, cocos2d::SEL_MenuHandler, GJTimedLevelType) override;
 
-    void createWeekdayLabel(const char* text, int idx);
-    void loadMonth();
-    void setupMonth();
+    void createWeekdayLabel(const char*, int);
+    void loadMonth(int, int, bool = false);
+    void loadSafe(bool = false);
+    void setupMonth(cocos2d::CCArray*);
 public:
     static BSCalendarPopup* create(CCObject*, cocos2d::SEL_MenuHandler, GJTimedLevelType);
 
-    void loadLevelsFinished(cocos2d::CCArray* levels, const char* key, int) override {
-        loadLevelsFinished(levels, key);
-    }
-    void loadLevelsFinished(cocos2d::CCArray*, const char*) override;
-    void loadLevelsFailed(const char* key, int) override {
-        loadLevelsFailed(key);
-    }
-    void loadLevelsFailed(const char*) override;
-    void setupPageInfo(gd::string, const char*) override {}
+    GJTimedLevelType getType() const { return m_type; }
+    int getYear() const { return m_year; }
+    int getMonth() const { return m_month; }
+    int getMinYear() const { return m_minYear; }
+    int getMaxYear() const { return m_maxYear; }
+    int getMinMonth() const { return m_minMonth; }
+    int getMaxMonth() const { return m_maxMonth; }
+
+    void loadLevelsFinished(cocos2d::CCArray*, const char*, int) override;
+    void loadLevelsFailed(const char*, int) override;
 
     ~BSCalendarPopup() override;
 };