+++ src/classes/BSHoverNode.cpp
@@ -1,16 +1,18 @@
 #include "BSHoverNode.hpp"
 #include <Geode/binding/ButtonSprite.hpp>
-#include <Geode/binding/GJGameLevel.hpp>
+#include <Geode/binding/GameManager.hpp>
 #include <Geode/binding/GameLevelManager.hpp>
 #include <Geode/binding/GameStatsManager.hpp>
+#include <Geode/binding/GJGameLevel.hpp>
+#include <Geode/binding/ProfilePage.hpp>
 #include <Geode/utils/cocos.hpp>
 #include <Geode/ui/Layout.hpp>
 
 using namespace geode::prelude;
 
-BSHoverNode* BSHoverNode::create(const SafeLevel& level, GJGameLevel* gameLevel, const std::function<void()>& callback) {
+BSHoverNode* BSHoverNode::create(const SafeLevel& level, GJGameLevel* gameLevel, HoverCallback callback) {
     auto ret = new BSHoverNode();
-    if (ret->init(level, gameLevel, callback)) {
+    if (ret->init(level, gameLevel, std::move(callback))) {
         ret->autorelease();
         return ret;
     }
@@ -18,25 +20,25 @@ BSHoverNode* BSHoverNode::create(const SafeLevel& level, GJGameLevel* gameLevel,
     return nullptr;
 }
 
-bool BSHoverNode::init(const SafeLevel& level, GJGameLevel* gameLevel, const std::function<void()>& callback) {
+bool BSHoverNode::init(const SafeLevel& level, GJGameLevel* gameLevel, HoverCallback callback) {
     if (!CCLayer::init()) return false;
 
     setAnchorPoint({ 0.5f, 0.5f });
-    setContentSize({ 80.0f, 70.0f });
+    setContentSize({ 80.0f, 50.0f });
     setID("BSHoverNode");
     ignoreAnchorPointForPosition(false);
 
     CCTouchDispatcher::get()->registerForcePrio(this, 2);
 
-    m_callback = callback;
+    m_callback = std::move(callback);
 
-    m_background = CCScale9Sprite::create("square02_001.png");
-    m_background->setContentSize({ 80.0f, 70.0f });
-    m_background->setPosition({ 40.0f, 35.0f });
-    m_background->setColor({ 0, 0, 0 });
-    m_background->setOpacity(150);
-    m_background->setID("background");
-    addChild(m_background);
+    auto background = CCScale9Sprite::create("square02_001.png");
+    background->setContentSize({ 80.0f, 50.0f });
+    background->setPosition({ 40.0f, 25.0f });
+    background->setColor({ 0, 0, 0 });
+    background->setOpacity(150);
+    background->setID("background");
+    addChild(background);
 
     auto timedName = "";
     switch (level.type) {
@@ -46,33 +48,50 @@ bool BSHoverNode::init(const SafeLevel& level, GJGameLevel* gameLevel, const std
         default: break;
     }
     auto dailyLabel = CCLabelBMFont::create(fmt::format("{} #{}", timedName, level.timelyID).c_str(), "goldFont.fnt");
-    dailyLabel->setPosition({ 40.0f, 65.0f });
+    dailyLabel->setPosition({ 40.0f, 45.0f });
     dailyLabel->setScale(0.3f);
     dailyLabel->setID("daily-label");
     addChild(dailyLabel);
 
+    auto viewMenu = CCMenu::create();
+    viewMenu->setPosition({ getContentWidth() / 2.0f, 0.0f });
+    viewMenu->setID("view-menu");
+    addChild(viewMenu);
+
+    auto closeButton = CCMenuItemExt::createSpriteExtraWithFrameName("GJ_closeBtn_001.png", 0.5f, [this](auto) { close(); });
+    closeButton->setPosition({ -40.0f, 50.0f });
+    closeButton->setID("close-button");
+    viewMenu->addChild(closeButton, 1);
+
     auto nameLabel = CCLabelBMFont::create(gameLevel->m_levelName.c_str(), "bigFont.fnt");
-    nameLabel->setPosition({ 40.0f, 55.0f });
     nameLabel->setScale(0.5f);
-    nameLabel->setID("name-label");
-    addChild(nameLabel);
+    auto nameButton = CCMenuItemExt::createSpriteExtra(nameLabel, [gameLevel](auto) { GameLevelManager::get()->gotoLevelPage(gameLevel); });
+    nameButton->setPosition({ 0.0f, 35.0f });
+    nameButton->setID("name-button");
+    viewMenu->addChild(nameButton);
 
-    auto creatorLabel = CCLabelBMFont::create(("by " + std::string(gameLevel->m_creatorName)).c_str(), "goldFont.fnt");
-    creatorLabel->setPosition({ 40.0f, 43.0f });
+    auto creatorLabel = CCLabelBMFont::create(fmt::format("by {}", GEODE_ANDROID(std::string)(gameLevel->m_creatorName)).c_str(), "goldFont.fnt");
     creatorLabel->setScale(0.4f);
-    creatorLabel->setID("creator-label");
-    addChild(creatorLabel);
+    auto creatorButton = CCMenuItemExt::createSpriteExtra(creatorLabel, [gameLevel](auto) {
+        if (gameLevel->m_accountID.value() > 0) {
+            GameManager::get()->setGameVariable("0047", true);
+            ProfilePage::create(gameLevel->m_accountID.value(), false)->show();
+        }
+    });
+    creatorButton->setPosition({ 0.0f, 23.0f });
+    creatorButton->setID("creator-button");
+    viewMenu->addChild(creatorButton);
 
-    setContentWidth(std::max(nameLabel->getScaledContentWidth() + 6.0f, std::max(creatorLabel->getScaledContentWidth() + 6.0f, 80.0f)));
-
-    m_background->setContentWidth(getContentWidth());
-    m_background->setPositionX(getContentWidth() / 2);
-    dailyLabel->setPositionX(getContentWidth() / 2);
-    nameLabel->setPositionX(getContentWidth() / 2);
-    creatorLabel->setPositionX(getContentWidth() / 2);
+    auto contentWidth = std::max({ nameLabel->getScaledContentWidth() + 6.0f, creatorLabel->getScaledContentWidth() + 6.0f, 80.0f });
+    setContentWidth(contentWidth);
+    background->setContentWidth(contentWidth);
+    background->setPositionX(contentWidth / 2.0f);
+    dailyLabel->setPositionX(contentWidth / 2.0f);
+    viewMenu->setPositionX(contentWidth / 2.0f);
+    closeButton->setPositionX(-contentWidth / 2.0f);
 
     auto starLayout = CCNode::create();
-    starLayout->setPosition({ getContentWidth() / 2, 30.0f });
+    starLayout->setPosition({ contentWidth / 2.0f, 10.0f });
     starLayout->setContentSize({ 80.0f, 15.0f });
     starLayout->setAnchorPoint({ 0.5f, 0.5f });
     starLayout->setLayout(RowLayout::create()->setGap(1.75f)->setAutoScale(false));
@@ -80,11 +99,11 @@ bool BSHoverNode::init(const SafeLevel& level, GJGameLevel* gameLevel, const std
     addChild(starLayout);
 
     auto gsm = GameStatsManager::get();
+    auto levelID = gameLevel->m_levelID.value();
+
     auto starsLabel = CCLabelBMFont::create(std::to_string(gameLevel->m_stars.value()).c_str(), "bigFont.fnt");
     starsLabel->setScale(0.4f);
-    auto levelID = gameLevel->m_levelID.value();
-    auto completedLevel = gsm->m_completedLevels->objectForKey(fmt::format("c_{}", levelID).c_str());
-    starsLabel->setColor(completedLevel ? ccColor3B { 255, 255, 50 } : ccColor3B { 255, 255, 255 });
+    starsLabel->setColor({ 255, 255, (uint8_t)(255 - gsm->hasCompletedOnlineLevel(levelID) * 205) });
     starsLabel->setID("stars-label");
     starLayout->addChild(starsLabel);
 
@@ -97,31 +116,16 @@ bool BSHoverNode::init(const SafeLevel& level, GJGameLevel* gameLevel, const std
         auto coinStr = fmt::format("{}_{}", levelID, i);
         auto hasCoin = gsm->hasUserCoin(coinStr.c_str()) || gsm->hasPendingUserCoin(coinStr.c_str());
         auto coinSprite = CCSprite::createWithSpriteFrameName("usercoin_small01_001.png");
-        if (coinsVerified) coinSprite->setColor(hasCoin ? ccColor3B { 255, 255, 255 } : ccColor3B { 165, 165, 165 });
-        else coinSprite->setColor(hasCoin ? ccColor3B { 255, 175, 75 } : ccColor3B { 165, 113, 48 });
+        coinSprite->setColor({
+            (uint8_t)(165 + hasCoin * 90),
+            (uint8_t)(113 + coinsVerified * 52 + hasCoin * (62 + coinsVerified * 28)),
+            (uint8_t)(48 + coinsVerified * 117 + hasCoin * (17 + coinsVerified * 73))
+        });
         coinSprite->setID(fmt::format("coin-sprite-{}", i).c_str());
         starLayout->addChild(coinSprite);
     }
 
     starLayout->updateLayout();
-
-    auto viewMenu = CCMenu::create();
-    viewMenu->setPosition({ getContentWidth() / 2, 12.0f });
-    viewMenu->setID("view-menu");
-    addChild(viewMenu);
-
-    auto viewSprite = ButtonSprite::create("View", "goldFont.fnt", "GJ_button_01.png", 0.8f);
-    viewSprite->setScale(0.7f);
-    auto viewButton = CCMenuItemExt::createSpriteExtra(viewSprite, [this, gameLevel](auto) {
-        GameLevelManager::get()->gotoLevelPage(gameLevel);
-    });
-    viewButton->setID("view-button");
-    viewMenu->addChild(viewButton);
-
-    auto closeButton = CCMenuItemExt::createSpriteExtraWithFrameName("GJ_closeBtn_001.png", 0.5f, [this] (auto) { close(); });
-    closeButton->setPosition(viewMenu->convertToNodeSpace(convertToWorldSpace({ 0.0f, 70.0f })));
-    closeButton->setID("close-button");
-    viewMenu->addChild(closeButton);
 
     setTouchEnabled(true);
     setKeypadEnabled(true);
@@ -145,18 +149,14 @@ void BSHoverNode::registerWithTouchDispatcher() {
     CCTouchDispatcher::get()->addTargetedDelegate(this, -500, true);
 }
 
-// Thanks Firee https://github.com/FireMario211/Object-Workshop/blob/v1.1.3/src/nodes/ExtPreviewBG.cpp#L40
 bool BSHoverNode::ccTouchBegan(CCTouch* touch, CCEvent* event) {
     if (!CCLayer::ccTouchBegan(touch, event)) return false;
 
-    auto touchPos = m_background->convertToNodeSpace(CCDirector::get()->convertToGL(touch->getLocationInView()));
-    auto boundingBox = m_background->boundingBox();
-    return CCRect {
-        boundingBox.getMinX(),
-        boundingBox.getMinY(),
-        boundingBox.getMaxX(),
-        boundingBox.getMaxY()
-    }.intersectsRect({ touchPos.x, touchPos.y, touchPos.x, touchPos.y });
+    auto position = convertToWorldSpace({ 0.0f, 0.0f });
+    auto& size = getContentSize();
+    auto touchPosition = touch->getLocation();
+    return position.x >= touchPosition.x && position.x + size.width <= touchPosition.x
+        && position.y >= touchPosition.y && position.y + size.height <= touchPosition.y;
 }
 
 BSHoverNode::~BSHoverNode() {