+++ changelog.md
@@ -1,11 +1,16 @@
 # Better Safe Changelog
+## v1.1.4 (2025-03-25)
+- Redesigned the hover previews to be more compact
+- Changed the background color of the month selection popup to match the daily/weekly/event level previews
+- Added particles to mythic rated levels
+
 ## v1.1.3 (2025-02-07)
 - Added node IDs to the popups
 - Updated the gallery to include the new event levels
 
 ## v1.1.2 (2024-11-20)
 - Changed the background color of the calendar to match the daily/weekly/event level previews
-- Fixed a bug where the refresh button would not load the levels properly ([#5](https://github.com/hiimjustin000/BetterSafe/issues/5))
+- Fixed a bug where the refresh button would not load the levels properly ([#5](https://github.com/hiimjasmine00/BetterSafe/issues/5))
 
 ## v1.1.1 (2024-11-19)
 - Fixed a bug where the event level would not properly fill in the current day if you were not British